ev-window: remove sidebar_page_main_widget_update_cb
This was introduced in 4f1722b0523879cd73ee936d483a735e163a6b54 for
bindings to work in the sidebar. This is no longer an issue, and
bindings keep working in the sidebar regardless of this.
Probably now with GTK3 the events are sent correctly to the focused
widgets, although it's hard to tell why this workaround is unnecessary
when it was introduced 18 years ago with little additional information
around it (like a bugzilla ticket).