libview: rework selection and annotation move with GtkGestureDrag
The logic is split from the callbacks of GtkGestureClick press and
release + GtkEventControllerMotion motion into its own
GtkGestureDrag (which is actually the gesture that takes care of
that). This allows to put a lot more logic local to those callbacks,
making several of the fields in the privates unnecessary. One big
advantage of this is that the global "pressed_button" can be removed
too, and it makes it a lot clearer what do each of the controllers
work for.

TODO:
 - Fix SHIFT selection, take from previous commit