search-context: Add results as they appear
We are waiting until the search job is finished to add this results.
On very large PDFs this makes search appear to be incredibly slow.
Instead add the results as they appear (like evince did). Move
the spinner to the right of the SearchBox (instead of having an
empty state). Otherwise all of the logic essentially remains the
same except the use of a GtkSortListModel to ensure proper sorting.