ci: override flatpak stable SDK and allow failure for release branches
The branch in our manifest should reflect the one the
branch is supposed to be shipped with. For once, this can
surface accidentally undeclared dependency bumps with
backported changes, and additionally makes sure it used
built and tested using the stylesheets, such as Adwaita's
for the targeted version in question.

The gnome-runtime-image used by the flatpak CI pipeline
only provides the master branch of the SDK and no stable
branch, so the simplest solution is to build on master in
CI and pray it works. If it does not, however, it should
be allowed to fail, in which case it should be made sure
locally, it is in fact due to the newer SDK version.

Backported caccce5a2b76e7e515da89a79429a95850cef20e

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>