libview: view: use search context's result selection
The search context now has a selection model for this
very purpose, so that PpsView does not need to keep this
field, this also enales us to use the result model's
notify::selected-item directly.

PpsSearchContext::result-activated is a pass-through of
selection-changed which always fires when the selected
position changes unless it is autoselected. This does
not change the behaviour of Papers currenty, enables the
removal of the pass-through signal, and allows for more
freedom with regards to sorting and autoselecting in
the future with PpsView still following autoselected
items.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>