libview: view: handle and react to search activation
The search context now has the concept of being active
and inactive. The view herewith makes use of it.

However, the pps_view_find_set_highlight_search no longer
makes sense, as the search context is supposed to handle
it. Thus, it will be removed in a subsequent commit,
after the shell side of this change is done.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>