shell: find-sidebar: replace binary search over all results with direct per-page API
While binary search was faster than the previous linear
search over all results, with the introduction of the
new per-page results API we can directly access the
result we want.

For the restart's linear search, we also iterate over
the much shorter list of pages between the current page
and the page of the last overall search result.

This is especially faster for when we know the last
result is before the current page, as then there is no
need for the search at all.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>