libview: annotation-window: ignore X11 deprecation
While we should definitely migrate away from X11-specific
code, these two hints do not hurt, are local, do not have
a corresponding Wayland flag, and will be easily removed
when we migrate to GTK5.

Hopefully, this window will instead be replaced with
on-view widgets long before the migration to GTK5 and
the subsequent removal of GDK's X11 backend.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>