libview,shell: connect callbacks to reload annot area when annot changes
This allows a small simplification on the shell, where we can avoid
calls to libview, and a performance optimization, as we avoid
reloading the whole document on annotation changed. This requires the
introduction of a new signal to the annotations context, since there
is otherwise no way to exactly know when is the context finished with
its loading and its safe to connect to callbacks to all the annotations!