libview: view: only react to rotation changes if document is set
Once a document is set, caches will need to be filled,
and the document measured, allocated and drawn. Those
will make use of the set rotation at that point in time.
There is nothing we would need to or could prepare here.

This fixes a critical that resulted from clearing the
selection without a document, and consequently no pixbuf
cache.

Fixes: 400e74d2a41d0144eafc97bf94bc2101f79e77ff ("libview: view: use zoom state of document model instead of local copy")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>