shell: window: get file's edit name only if info is known to exist
Per its documentation, it is an error to be called when
the info is not present.

G_FILE_ATTRIBUTE_STANDARD_EDIT_NAME is not guaranteed to
be set, and its purpose is it is to be used as the
initial value for rename operations with
Gio.File.set_display_name. As such, the display name is a
sensible fallback.

Link: https://docs.gtk.org/gio/method.FileInfo.get_edit_name.html
Fixes: fc9370419720915953c5e97fe44c561ea78bb6fa ("shell: Use display and edit name for filename")
Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/364

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>