Revert "libview: remove PpsSelectionStyle from SelectionInfo"
This reverts commit 90e0dd79933d0f7be12bf7f40aae09e55f5aa08a.

The removal of SelectionInfo->style blocks the use of
this value in the selection drag gesture, where it is
required for the common double tap and drag gesture.

Besides the revert, I also add a comment explaining the
revert, because it is not always obvious, why this is
required here.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/369

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>