meson: Fix the logic of determining rust_target
The current logic of pps_debug is broken and more likely a switch to
determine whether we should enable debug symbol and not the optimization
level. For the rust_target, we use it more likely for the optimization
level.

There is a simple table for the optimization level used by various
build_type:

https://mesonbuild.com/Builtin-options.html#details-for-buildtype

So the condition get_option('optimization') == '0' holds when build_type
== 'debug'.

We can simply set the buildtype to 'debug' to make cargo build
unoptimized code.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>