libview: thaw notify signals in the document model when the document is being changed
Otherwise, the internal reference priv->document is modified, then the page changed signal
is emitted but consumers of the document model may not know that the document has changed and
therefore may consider a wrong number of pages. Therefore, it is better to first change the model
page to a valid one and only then emit all notify signals.

Related to https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/353