libview: view: fix initial size allocation for free sizing mode
When the document is loaded in free sizing mode the
measure call happens once, and the subsequent call to
allocate sets priv->pending_resize to FALSE. And as
resizing the widget should not resize the document, there
is not subsequent call to measure the document.

However, that happens too early for the scale to be set,
so that the adjustment's values are too large. Instead,
we should rely on pps_view_update_adjustment_values being
idempotent.

Fixes: efb7101945a873e930a5098ad0e03e732c481e89 ("libview: view: use GTK's allocation queue instead of custom solution")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>