libview: view: let pps_view_handle_cursor_over_xy also set cursors for link
pps_view_handle_cursor_over_xy now handles every cursor
change and triggers handle_cursor_over_link which is now
named handle_link_preview with a narrower scope.

This allows to split out handling link previews of our
hot paths, in which it is currently running, as well as
making sure it does not running during paint phase, in
which it is a bug to alter the widget tree.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>