thumbnailer: port to Rust
Ports the thumbnailer to Rust. This includes simplifying how files are loaded,
skipping to create a tempfile, as that seems to be no longer necessary. As well
as dropping the `no-limit` flag, as it was introduced in 2011 to keep the old
buggy behavior of indefinitely blocking on a thumbnail generation.
With the changes in software since, that is no longer necessary any,
as if a document thumbnail cannot be created within 15 it is better to fail.
Additionally, the code for calling the thumbnailers does not support setting
the `no-limit` flag.