libview: view: split up scrolling to page position
scrolling to page position is used to scroll to a page,
e.g. when clicking on a link, but also to scroll to a
specific position within that page, as well as scrolling
back to the current page, after a resize.

These are completely different use cases of the same
code path and lead to the current situation of storing
a point, a boolean, and an enum value that together
influence the complicated adjustment value change
functions' behaviour.

Instead, it makes sense to split it up by these use cases.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/362

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>