libview: view: let queued allocation be handled after future realisation
If on an unrealised widget a reallocation is queued, it
will only be done after its next realisation. But at that
point we want that to happen, because we wanted to signal
that our current allocation is no longer valid.

Thus, checking for whether PpsView was realised was not
correct. Additionally, the commented out code is already
handled again at allocation time since the commit
1148ac3c14b3a97da6fda42772c785763305fe31 ("libview: view: always handle cursor after allocation").

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>