libview: stop using annot_mapping from PpsPageCache
Using the annot_mapping has no real benefit but a small optimization
in "show_annots_border", which is only active during debugging. So it
simplifies the code, and has no actual impact.

Helps #382

Also note, we have to set the context before setup_default in the
shell, as otherwise there is no model when doing annotations.