pdf: do not cache annotations per-page in the backend
Since the introduction of the annotations context, and the removal of
the annotations from the page cache, this has become an unnecessary
optimization. What holds the cache of the annotations is the context,
and that results in the "get_annotations" basically only be called
once per document load. And re-loading invalidates the cache anyway!
If we ever need an optimization for this, it should be done in the
annotations context similarly to how it was done in the search context
in d155a02a278b3b537cb0c039d0d9f90a1c741c0b