shell: stop cloning not transferred references to GObjects
In all cases the clone is unnecessary. In some cases it
is only that as the function's lifetime is short, but in
most cases these references are passed to external
libraries with transfer none.

Regardless of whether every single one here is a memory
leak, not having a new reference helps with not needing
to argue about when it leaves scope.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>