Do not load documents from pps_document_factory functions
These functions are designed to provide the raw class, and not
actually load documents. The behavior is not documented and can cause
issues, like in the nautilus extension, where the document is then
loaded again! This API is also synchronous and can lock the main
thread of apps. So its a much worse alternative to PpsJobLoad