Load cache after a successful PpsJobLoad instead of in libdocument
This allows to cleanup the API for pps_job_load, and to more clearly
specify which helper functions from pps_document_get* require a cache
to be setup. This flags in the API were introduced in
https://bugzilla.gnome.org/show_bug.cgi?id=780210 as an optimization
for the thumbnailer. However, having the flags means that every
consumer of the API at all levels (PpsJobLoad and below) needed to be
aware of this flag. The changes in this commit mean that most
consumers that "just want a document" and will be using PpsJobs do not
have to worry at all about the flags, but get the document they
want. Those that care about performance can use the lower-level
PpsDocument-based API, and avoid setting the caches.