shell: find-sidebar: Only allow tabbing through a single list item
Navigating through the find-sidebar page can become cumbersome when
the user has to go through tons of list items with their keyboard.
The tabbing behavior also works differently as opposed to using
arrow buttons. The former focuses on the item, whereas the latter
focuses and selects the item.

This patch addresses that behavior by allowing the user to tab
through a single item. When tabbing through the list, the
selected item is focused and selected, as opposed to the
current implementation where tabbing through the list focuses
and scrolls all the way to the first item in the list.