libview: fix pps_job_run signature and documentation
The documentation is wrong, the return value was once used when we had
PpsJob subclasses of "idle" type, where the return value would tell
the idle whether to re-run or remove it. Now it's no longer the case,
and we just unconditionally return FALSE, making the whole return value
unnecessary.