libdocument/pps-poppler: Do not leak the signing data
We were leaking the signing data, and for a "good" reason, since poppler
upstream is buggy and does not make a copy of the data when passes it to
the task thread.

However, in the mean time it's still better to workaround the issue than
just accept the leak, so keep the data bound to the signing task until
we're done with the signature.

Keeping the code in a way that this commit can be just reverted
(minus the autopointer side) when we require a poppler version that is
new enough.

See:
  https://gitlab.freedesktop.org/poppler/poppler/-/merge_requests/1803
(cherry picked from commit 5d68d8f7cc6485c46ee244036d840bb9d9aa12cd)