libview/pps-view: Clamp the signing selection to the selected page area
When signing a document and the selection area bounds are outside the
document area we are currently crashing.

This happens because in such case the view marks we're getting are
null (one or both of the two), and we end up dereferncing NULL start or
stop marks.

So if this happens, try instead to figure out in what page we are by
looking for a valid mark (if any otherwise we go best efforts and pick
the center), and eventually clamp the selection area to such page
extents, so that we are not only crashing but the signature is always
fully visible.

In case the selection happens fully outside a document page we're now
instead showing the "selection too small" alert dialog.