libview/pps-view: Do not replace the sign cursor on drag-less movements
When starting the sign rectangle selection we should show the crossair
cursor, but this gets replaced at the first movement because at that
moment the gesture is not active yet (we're not dragging!).

So, instead of checking whether we're dragging or not, let's check
whether the gesture propagation is monitored.