libview/pps-view: Ignore the scroll offset when drawing the sign area
When doing a visible signature on a page that requires
scrolling, the signature rectangle preview wasn't shown,
while the actual signed form was placed in the right
spot.

This was because we used the same scrolling offset
position for both drawing and signing, while we should
not use any offset when drawing.

Co-Authored-By: Markus Göllnitz <camelcasenick@bewares.it>

Cherry-picked f9c7855fda231230cd1b2934fcc9a90e9d5f84ac

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>