libdocument: pdf: connect signals to detect annotation changes
This fixes a potential race condition modifying annotations. Every
consumer first modifies an annotation, the signals associated with
that annotation fire, and then the annotation is saved. This is a
mess. Instead, we should create the annotation, connect the associated
modification callbacks, and then any other user callbacks. This means,
we are going to remove the "save_annotation" in future commits. Work
to do:

- How to mark the document as modified. We need the doc in the
callbacks
- How to fix markup type changes. We also need the doc in the
callbacks.
- Reusing the callbacks on "add annotation" is nice, but cumbersome
with the pspec in there