libdocument: Free the first error when trying to open an unsupported file
This patch fixes the following warnings when opening an unsupported file.

```console
(papers:627877): GLib-WARNING **: 12:41:49.514: GError set over the top of a previous GError or uninitialized memory.
This indicates a bug in someone's code. You must ensure an error is NULL before it's set.
The overwriting error message was: File type Plain text document (text/plain) is not supported
```