help: remove default-settings page
The documentation is outdated and we made the conscious decision to
not support this pattern in 64ef4464fd121673a514a9d51cd399f5a95aacb0.
The idea is that settings should be auto-saved on close and be well
supported for documents that people don't have to think about it.

Fixes #443

(cherry picked from commit 2a39ec541515b5532faf6d0eb7f5a65b815be59b)