libdocument: add get_page_size_uncached
This is an alternative to the regular get_page_size, that requires the
cache to be loaded. This is necessary because from the thumbnailer we
can no longer call the C class virtual function as we were doing since
f6d001ad51b659c8bf9c10d4466182f4ccfcad95. It will be used in the next
commits.