libview: move handling of annot drag from PpsView to PpsViewPage
To make it possible, create a helper function in the annotations
context that allows us to fetch an annotation given a certain
PpsDocumentPoint. This helps move the logic out of PpsView, and use it
from both the view and the view page.

Unfortunately, this change also duplicates a bit of logic regarding
transforming points in coordinate systems until we are able to move
more things to the page.