libview: view: handle scrolling callback through adjustments
The event controller is not meant for passive listening
of scrolling events. Its events are captured by the
scrolled window around the view. To make this improper
use clear, this renames the function to a single purpose
handler for zooming.

The other use case of removing the link preview popover
is part of the cursor hovering over different content
when it is scrolled. Thus, adjustment changes are the
proper signal to listen to here, where complete handle
cursor updates are warranted as scrolling away from a
link is not the only possibility. That is, however,
already covered at re-allocation.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/465

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>