libview: view: unset preview link properly
Hovering over a link, letting the preview close, and
hovering the same link as before makes the check that
the new link and the old link must differ fail.

If you are not hovering over a link, the current preview
link, arguably is NULL. And setting it to this value
fixes said issue.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/466

Cherry-picked b0606f128096be9347b2f607a067bb6cdca58470

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>