meson: Move the config.rs.in to the rust/ directory
After adding the Rust-based thumbnailer, the meson devenv is messed up.
That is, the environment used to build the cargo project is inconsistent
with the environment setup by the `meson devenv`. So the following two
way to build the project are conflicted and can't share the building
cache:

1. ninja -C builddir
2. meson devenv -C builddir && ninja

We solve this by splitting a common cargo environment. And passing the
resources file through an environment variable. This also enable us to
use config.rs.in in multiple Rust-based components.

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>