shell: application: only terminate foreground tasks on quit
Currently, the app.quit action, unlike win.close, will
exit the application without prompting the user to save
documents with unsaved changes, or without asking whether
to continue printing etc.

This also is a basis for other features such as
continuing printing in background before exiting by
simply explicitly holding the application for print
actions. That would eliminate the need to prompt the user
to determine whether to continue to do so as done now.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>