libview+shell: for search use selection model and listview activation directly
This depends on GTK splitting controls for activation and
selection in listviews. Those are two signals, but before
the ability to have both at single click without
selection on hover, this was not possible.

This allows us to remove our custom autoselection and
activation workaround, where we used the selection change
as an indicator for activation, unless called through our
autoselection API.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>