shell: document-view: move annotation context menu initialisation to UI file
There is no need to add the item at runtime. Instead, we
should only hide it, depending on whether annotations can
be added and modified in a document or not.

For that, the custom widget needs to be added with the
custom type set in the model. And due to Gtk.Builder
reading the file from top to bottom, the order matters.
Hence, the menu needs to be parsed before the child can
be added. Thus, the menus are moved to the top.

Depends on: https://gitlab.gnome.org/GNOME/gtk/-/issues/7698

Closes: https://gitlab.gnome.org/GNOME/papers/-/issues/490

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>