shell: sidebar-attachments: fix unselecting gesture
We want to unselect everything when a user clicks
somewhere outside the list. There, however, still is the
list view and gtk_widget_pick reports it (and would only
report NULL/None for coordinates completely outside the
widget, which cannot happen with event handler provided
coordinates, as they are within it by definition).

Fixes: 4bba1ea146943786eb98fea1db516e8cc0b43c4f ("shell: Port PpsSidebarAttachments to Rust")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>