shell: sidebar-thumbnails: replace suppressed selection signals with activation
This surfaces the fact that set_current_page is set twice
upon start-up – once through map and once through the
then selection signal handler.

This is now fixed, but requires an idle callback to allow
for the grid to be measured and allocated, so that
scrolling to the initial page thumbnail works correctly.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>