shell: add Edit Note option to the sidebar annot menu
This implementation adds an Edit Note entry to the sidebar annotation context menu, allowing users to edit annotation notes directly without having to search for them in the document view. To improve usability, it also ensures that the annotation editor window is properly presented and focused when triggered from the sidebar. This mitigates https://gitlab.gnome.org/GNOME/papers/-/issues/399.