libview,shell: provide a single item on menu popup
Providing multiple items on popup menu selection increases the
complexity remarkably (see
https://gitlab.gnome.org/GNOME/papers/-/merge_requests/674). In case
of overlapping entities, then we have to do our best guess, and chose
based on our best options. Same as if right clicking on something that
has a link and an annotation should not be expected to both go to the
link, and open the annotation, we should not have to deal with
multiple things at the same time. Therefore, we establish a priority
of: annot, link, image. Further heuristics could be also developed.

This partially reverts 5ef0d418d651705e098d2a1a8495e86bf7c0f538
developed from https://bugzilla.gnome.org/show_bug.cgi?id=609497 In
its original form it complains about an image and a link overlapping:

> you may note that, while the entire image is clickable as a link, in
> some cases, the visual representation of the link text is the only
> part that will produce the link context entries (such as "Copy Link"),
> and in the remaining cases, it doesn't seem to matter where you click,
> you won't get the link options.

Given the whole history of this app, that seems like a completely
reasonable behavior.

The shell is simply adjusted accordingly.