shell: Add toast notification for annotation deletion
Add a toast notification that appears when an annotation is deleted,
displaying “Annotation deleted” with an “Undo” button. The toast
appears at the bottom of the window for 5 seconds and allows users
to quickly restore accidentally deleted annotations by clicking the
Undo button or using Ctrl+Z or Alt+U.

This improves feedback for deletion actions and helps users discover
that deleted annotations can be restored.

Also add rust bindings for the undo context.

Closes https://gitlab.gnome.org/GNOME/papers/-/issues/483