libview: auto-rotate pages by default when printing
This was the default when there was a custom widget in the print dialog (the checkbox
was checked by default and used to set print->autorotate). But as
none of the custom widget functions are called in flatpak/recent gtk versions,
autorotate is never modified from its default false value.

Closes https://gitlab.gnome.org/GNOME/papers/-/issues/546